Fix segfault related to homegrown tilde replacement.
I want to test the tilde expantion code on a musl libc based system
before I am happy with this commit.

@troglobit: Have you run the homegrown tilde solution on musl libc
based systems yet? I cannot seem to google my way to an answer as to
why I need to work on a copy of arg when calling memmove() atm. :/

Signed-off-by: Magnus Malm <magnusmalm@gmail.com>